refactor scatterv
